check if the tree is perfect

+++ 16-binary_tree_is_perfect.c
@@ -35,7 +35,7 @@ int binary_tree_is_perfect(const binary_tree_t *tree)
  */
 size_t binary_tree_height(const binary_tree_t *tree)
 {
-	int leftHeight, rightHeight;
+	size_t leftHeight, rightHeight;
 
 	if (!tree)
 		return (0);
@@ -43,4 +43,5 @@ size_t binary_tree_height(const binary_tree_t *tree)
 	leftHeight = binary_tree_height(tree->left);
 	rightHeight = binary_tree_height(tree->right);
 
-	return (1 + (leftHeight > rightHeight ? leftHeight : rightHeight));}
+	return (1 + (leftHeight > rightHeight ? leftHeight : rightHeight));
+}
